[DCA-48] More appropriate CIDR range (#13)
There are a few problems with using the default CIDR range 10.0.0.0/16

* An AWS best practice is to setup every VPC in an organization
with a unique CIDR range.  Using the default will not be unique.

* The 10.0.0.0/16 CIDR range is too large thus reducing the number of
non-overlapping CIDRs that we can allocate in our organization.
 we want to setup with a more appropriate sized CIDR range.
The /24 is a smaller range which will allow us to have more unique
CIDRs to assign to VPCs.