[DCA-49] Fix the secret ID (#21)
The secret ID was incorrectly set to contain the env dictionary
which cause the deployment to fail.  The fix is to change it
to reference the context name which is just a string.

we are not creating any secrets therefore the create_secrets
function was renamed to make more sense.  Also there seems
to be no secret id at this time, only a secret name.