Refactor to deploy multiple envs to same aws account
*Note:* This is a 2nd attempt at commit 4df55d67 which was reverted
in 9ad49bc6 because it failed due to passing an invalid ID to the
ApplicationLoadBalancedFargateService function.

This is a refactor to allow deployments of the same app in
multiple environments (dev/test/stage/prod/etc..) to the
same AWS account. Since cloudformation stack names need to
be unique we append the environment context to the stack name
to uniquely identify the infra to the environment.  The
result of this refactor will make the CDK deploy stacks
with names like..  `dca-dev-DockerFargateStack` or
`dca-prod-DockerFargateStack` depending on which environment
is passed in for the context (i.e. cdk synth --context env=prod)

* Add the environment context to the cloudformation stack names
* Convert to fstrings
* Make vpc_stack get STACK_NAME_PREFIX from cdk context